Allow specifying output directories to build
Since 4dfc38d9 (Allow building only specific outputs,
2017-08-03) ngkbuild allowed users to specify which output files to only
build.

For the user's convenience, extend this functionality so that users can
also specify which output directories to only build. This is a short
hand for "only build output files within that directory".

A "directory" is distinguished from a "file" by: a directory path ends
with a path separator (e.g. "/a/b/"), while a file does not (e.g.
"/a/b")

Nexganku-Commit: 1192d72b9d59492bbd5887cc1d221e39fd48334c